+++ R/CA_plots.R
@@ -2,17 +2,27 @@
 
 ## Variables plot reactive data
 ## Not exported
-CA_var_data <- function(res, xax = 1, yax = 2, var_sup = TRUE, var_hide = "None",
+CA_var_data <- function(res, xax = 1, yax = 2, 
+                        lev_sup = TRUE, var_sup = TRUE,  var_sup_choice = NULL, 
+                        var_hide = "None",
                         var_lab_min_contrib = 0) {
     tmp_x <- res$vars %>% 
         filter(Axis == xax) %>%
-        select_("Level", "Position", "Type", "Class", "Coord", "Contrib", "Cos2", "Count")
+        select("Level", "Position", "Type", "Class", "Coord", "Contrib", "Cos2", "Count")
     tmp_y <- res$vars %>% 
         filter(Axis == yax) %>%
-        select_("Level", "Position", "Type", "Class", "Coord", "Contrib", "Cos2", "Count")
-    if (!var_sup) {
-        tmp_x <- tmp_x %>% filter(Type == 'Active')
-        tmp_y <- tmp_y %>% filter(Type == 'Active')
+        select("Level", "Position", "Type", "Class", "Coord", "Contrib", "Cos2", "Count")
+    if (!(var_sup) || is.null(var_sup_choice)) {
+        tmp_x <- tmp_x %>% filter(Type != 'Supplementary variable')
+        tmp_y <- tmp_y %>% filter(Type != 'Supplementary variable')
+    }
+    if (var_sup && !is.null(var_sup_choice)) {
+        tmp_x <- tmp_x %>% filter(Type != 'Supplementary variable' | Level %in% var_sup_choice)
+        tmp_y <- tmp_y %>% filter(Type != 'Supplementary variable' | Level %in% var_sup_choice)
+    }
+    if (!lev_sup) {
+        tmp_x <- tmp_x %>% filter(Type != 'Supplementary level')
+        tmp_y <- tmp_y %>% filter(Type != 'Supplementary level')
     }
     if (var_hide != "None") {
         tmp_x <- tmp_x %>% filter(Position != var_hide)
@@ -53,7 +63,9 @@ CA_var_data <- function(res, xax = 1, yax = 2, var_sup = TRUE, var_hide = "None"
 ##' @param res Result of prepare_results() call
 ##' @param xax Horizontal axis number
 ##' @param yax Vertical axis number
+##' @param lev_sup TRUE to display supplementary levels
 ##' @param var_sup TRUE to display supplementary variables
+##' @param var_sup_choice list of supplementary variables to display
 ##' @param var_hide elements to hide (rows or columns)
 ##' @param var_lab_min_contrib Contribution threshold to display points labels
 ##' @param point_size base point size
@@ -65,9 +77,12 @@ CA_var_data <- function(res, xax = 1, yax = 2, var_sup = TRUE, var_hide = "None"
 ##' @param in_explor wether the plot is to be displayed in the \code{explor} interface
 ##' @param ... Other arguments passed to scatterD3
 ##'
-##' @author Julien Barnier <julien.barnier@@ens-lyon.fr>
 ##' @export
-CA_var_plot <- function(res, xax = 1, yax = 2, var_sup = TRUE, var_hide = "None",
+CA_var_plot <- function(res, xax = 1, yax = 2, 
+                        lev_sup = TRUE,
+                        var_sup = TRUE, 
+                        var_sup_choice = NULL,
+                        var_hide = "None",
                         var_lab_min_contrib = 0,
                         point_size = 64,
                         col_var = NULL,
@@ -85,7 +100,7 @@ CA_var_plot <- function(res, xax = 1, yax = 2, var_sup = TRUE, var_hide = "None"
     lasso_callback <- if(in_explor) explor_multi_lasso_callback() else NULL
     zoom_callback <- if(in_explor) explor_multi_zoom_callback(type = "var") else NULL
 
-    var_data <- CA_var_data(res, xax, yax, var_sup, var_hide, var_lab_min_contrib)
+    var_data <- CA_var_data(res, xax, yax, lev_sup, var_sup, var_sup_choice, var_hide, var_lab_min_contrib)
 
     scatterD3::scatterD3(
                    x = var_data[, "Coord.x"],