+++ R/MCA_plots.R
@@ -2,17 +2,27 @@
 
 ## Variables plot reactive data
 ## Not exported
-MCA_var_data <- function(res, xax = 1, yax = 2, var_sup = TRUE, var_lab_min_contrib = 0) {
+MCA_var_data <- function(res, xax = 1, yax = 2, var_sup = TRUE, var_sup_choice = NULL,
+    var_lab_min_contrib = 0, labels_prepend_var = FALSE) {
+    
     tmp_x <- res$vars %>%
         arrange(Axis, Type, Variable) %>%
         filter(Axis == xax) %>%
-        select_("Variable", "Level", "Type", "Class", "Coord", "Contrib", "Cos2", "Count")
+        select("Variable", "Level", "Type", "Class", "Coord", "Contrib", "Cos2", "Count")
     tmp_y <- res$vars %>% 
         filter(Axis == yax) %>%
-        select_("Variable", "Level", "Type", "Class", "Coord", "Contrib", "Cos2", "Count")
-    if (!(var_sup)) {
+        select("Variable", "Level", "Type", "Class", "Coord", "Contrib", "Cos2", "Count")
+    if (!(var_sup) || is.null(var_sup_choice)) {
         tmp_x <- tmp_x %>% filter(Type == 'Active')
         tmp_y <- tmp_y %>% filter(Type == 'Active')
+    }
+    if (var_sup && !is.null(var_sup_choice)) {
+        tmp_x <- tmp_x %>% filter(Type == 'Active' | Variable %in% var_sup_choice)
+        tmp_y <- tmp_y %>% filter(Type == 'Active' | Variable %in% var_sup_choice)
+    }
+    if (labels_prepend_var) {
+        tmp_x$Level <- paste(tmp_x$Variable, "-", tmp_x$Level)
+        tmp_y$Level <- paste(tmp_y$Variable, "-", tmp_y$Level)
     }
     tmp <- tmp_x %>%
         left_join(tmp_y, by = c("Variable", "Level", "Type", "Class", "Count")) %>%
@@ -50,7 +60,9 @@ MCA_var_data <- function(res, xax = 1, yax = 2, var_sup = TRUE, var_lab_min_cont
 ##' @param xax Horizontal axis number
 ##' @param yax Vertical axis number
 ##' @param var_sup TRUE to display supplementary variables
+##' @param var_sup_choice list of supplementary variables to display
 ##' @param var_lab_min_contrib Contribution threshold to display points labels
+##' @param labels_prepend_var if TRUE, prepend variable names to labels
 ##' @param point_size base point size
 ##' @param col_var name of the variable for points color
 ##' @param symbol_var name of the variable for points symbol
@@ -60,10 +72,13 @@ MCA_var_data <- function(res, xax = 1, yax = 2, var_sup = TRUE, var_lab_min_cont
 ##' @param in_explor wether the plot is to be displayed in the \code{explor} interface
 ##' @param ... Other arguments passed to scatterD3
 ##'
-##' @author Julien Barnier <julien.barnier@@ens-lyon.fr>
 ##' @export
-MCA_var_plot <- function(res, xax = 1, yax = 2, var_sup = TRUE, var_lab_min_contrib = 0,
+MCA_var_plot <- function(res, xax = 1, yax = 2, 
+                         var_sup = TRUE, 
+                         var_sup_choice = NULL,
+                         var_lab_min_contrib = 0,
                          point_size = 64,
+                         labels_prepend_var = FALSE,
                          col_var = NULL,
                          symbol_var = NULL,
                          size_var = NULL,
@@ -79,7 +94,7 @@ MCA_var_plot <- function(res, xax = 1, yax = 2, var_sup = TRUE, var_lab_min_cont
     lasso_callback <- if(in_explor) explor_multi_lasso_callback() else NULL
     zoom_callback <- if(in_explor) explor_multi_zoom_callback(type = "var") else NULL
     
-    var_data <- MCA_var_data(res, xax, yax, var_sup, var_lab_min_contrib)
+    var_data <- MCA_var_data(res, xax, yax, var_sup, var_sup_choice, var_lab_min_contrib, labels_prepend_var)
     
     scatterD3::scatterD3(
                    x = var_data[, "Coord.x"],
@@ -141,7 +156,7 @@ MCA_ind_data <- function(res, xax = 1, yax = 2, ind_sup, col_var = NULL, opacity
                Lab = ifelse(Contrib >= as.numeric(ind_lab_min_contrib) | 
                               (is.na(Contrib) & as.numeric(ind_lab_min_contrib) == 0), Name, ""))
     if (!(is.null(col_var) || col_var %in% c("None", "Type"))) {
-        tmp_data <- res$quali_data %>% select_("Name", col_var)
+        tmp_data <- res$quali_data %>% select("Name", col_var)
         tmp <- tmp %>%
             left_join(tmp_data, by = "Name")
     }
@@ -167,7 +182,6 @@ MCA_ind_data <- function(res, xax = 1, yax = 2, ind_sup, col_var = NULL, opacity
 ##' @param in_explor wether the plot is to be displayed in the \code{explor} interface
 ##' @param ... Other arguments passed to scatterD3
 ##'
-##' @author Julien Barnier <julien.barnier@@ens-lyon.fr>
 ##' @export
 MCA_ind_plot <- function(res, xax = 1, yax = 2, ind_sup = TRUE, ind_lab_min_contrib = 0,
                          lab_var = NULL,
@@ -243,11 +257,15 @@ MCA_bi_data <- function(res, settings) {
     bi_data$Contrib[!ind] <- NA
     
     # Colors
-    if (settings$col_var == "Variable") {
+    if (!is.null(settings$col_var) && settings$col_var == "Variable") {
         bi_data$color <- bi_data$Variable
         bi_data$color[ind] <- ""
     } else {
-        bi_data$color <- bi_data[,settings$col_var]
+        if(is.null(settings$col_var)) {
+            bi_data$color <- NULL   
+        } else {
+            bi_data$color <- bi_data[,settings$col_var]
+        }
     }
 
     bi_data
@@ -275,7 +293,6 @@ MCA_bi_data <- function(res, settings) {
 ##' @param in_explor wether the plot is to be displayed in the \code{explor} interface
 ##' @param ... Other arguments passed to scatterD3
 ##'
-##' @author Julien Barnier <julien.barnier@@ens-lyon.fr>
 ##' @export
 ##' @importFrom RColorBrewer brewer.pal
 
@@ -306,7 +323,7 @@ MCA_biplot <- function(res, xax = 1, yax = 2,
     bi_data <- MCA_bi_data(res, settings)
     
     colors <- NULL
-    if (col_var == "Variable") {
+    if (!is.null(col_var) && col_var == "Variable") {
         n_colors <- nlevels(bi_data$color)
         if (n_colors <= 11) {
             colors <- c("#1f77b4", "#ff7f0e", "#2ca02c", "#d62728", "#9467bd", "#8c564b", "#e377c2", "#7f7f7f", "#bcbd22", "#17becf")