+++ R/PCA_plots.R
@@ -2,16 +2,21 @@
 
 ## Variables plot reactive data
 ## Not exported
-PCA_var_data <- function(res, xax = 1, yax = 2, var_sup = TRUE, var_lab_min_contrib = 0) {
+PCA_var_data <- function(res, xax = 1, yax = 2, var_sup = TRUE, 
+    var_sup_choice = NULL, var_lab_min_contrib = 0) {
     tmp_x <- res$vars %>% 
         filter(Axis == xax) %>%
-        select_("Variable", "Level", "Type", "Class", "Coord", "Contrib", "Cos2")
+        select("Variable", "Level", "Type", "Class", "Coord", "Contrib", "Cos2")
     tmp_y <- res$vars %>% 
         filter(Axis == yax) %>%
-        select_("Variable", "Level", "Type", "Class", "Coord", "Contrib", "Cos2")
-    if (!(var_sup)) {
+        select("Variable", "Level", "Type", "Class", "Coord", "Contrib", "Cos2")
+    if (!(var_sup) || is.null(var_sup_choice)) {
         tmp_x <- tmp_x %>% filter(Type == 'Active')
         tmp_y <- tmp_y %>% filter(Type == 'Active')
+    }
+    if (var_sup && !is.null(var_sup_choice)) {
+        tmp_x <- tmp_x %>% filter(Type == 'Active' | Variable %in% var_sup_choice)
+        tmp_y <- tmp_y %>% filter(Type == 'Active' | Variable %in% var_sup_choice)
     }
 
     tmp <- tmp_x %>%
@@ -51,6 +56,7 @@ PCA_var_data <- function(res, xax = 1, yax = 2, var_sup = TRUE, var_lab_min_cont
 ##' @param xax Horizontal axis number
 ##' @param yax Vertical axis number
 ##' @param var_sup TRUE to display supplementary variables
+##' @param var_sup_choice list of supplementary variables to display
 ##' @param var_lab_min_contrib Contribution threshold to display points labels
 ##' @param col_var name of the variable for points color
 ##' @param size_var name of the variable for points size
@@ -61,9 +67,10 @@ PCA_var_data <- function(res, xax = 1, yax = 2, var_sup = TRUE, var_lab_min_cont
 ##' @param in_explor wether the plot is to be displayed in the \code{explor} interface
 ##' @param ... Other arguments passed to scatterD3
 ##'
-##' @author Julien Barnier <julien.barnier@@ens-lyon.fr>
 ##' @export
-PCA_var_plot <- function(res, xax = 1, yax = 2, var_sup = TRUE, var_lab_min_contrib = 0,
+PCA_var_plot <- function(res, xax = 1, yax = 2, var_sup = TRUE, 
+                         var_sup_choice = NULL,
+                         var_lab_min_contrib = 0,
                          scale_unit = FALSE,
                          col_var = NULL,
                          size_var = NULL,
@@ -84,7 +91,7 @@ PCA_var_plot <- function(res, xax = 1, yax = 2, var_sup = TRUE, var_lab_min_cont
     if (is.null(xlim) && scale_unit && !has_quali_sup_vars) xlim <- c(-1.1, 1.1)
     if (is.null(ylim) && scale_unit && !has_quali_sup_vars) ylim <- c(-1.1, 1.1)
     
-    var_data <- PCA_var_data(res, xax, yax, var_sup, var_lab_min_contrib)
+    var_data <- PCA_var_data(res, xax, yax, var_sup, var_sup_choice, var_lab_min_contrib)
     
     scatterD3::scatterD3(
                    x = var_data[, "Coord.x"],
@@ -141,7 +148,7 @@ PCA_ind_data <- function(res, xax = 1, yax = 2, ind_sup = TRUE, col_var = NULL, 
                Lab = ifelse(Contrib >= as.numeric(ind_lab_min_contrib) | 
                               (is.na(Contrib) & as.numeric(ind_lab_min_contrib) == 0), Name, ""))
     if (!(is.null(col_var) || col_var %in% c("None", "Type"))) {
-        tmp_data <- res$quali_data %>% select_("Name", col_var)
+        tmp_data <- res$quali_data %>% select("Name", col_var)
         tmp <- tmp %>%
             left_join(tmp_data, by = "Name")
     }
@@ -167,7 +174,6 @@ PCA_ind_data <- function(res, xax = 1, yax = 2, ind_sup = TRUE, col_var = NULL, 
 ##' @param in_explor wether the plot is to be displayed in the \code{explor} interface
 ##' @param ... Other arguments passed to scatterD3
 ##'
-##' @author Julien Barnier <julien.barnier@@ens-lyon.fr>
 ##' @export
 PCA_ind_plot <- function(res, xax = 1, yax = 2, ind_sup = TRUE, ind_lab_min_contrib = 0,
                          col_var = NULL,