+++ R/explor_multi_MCA.R
@@ -128,11 +128,11 @@ explor.mca <- function(obj) {
 ##' ind_sup <- banque[1:100, -(19:21)]
 ##' var_sup <- banque[-(1:100),19:21]
 ##' acm <- dudi.acm(d, scannf = FALSE, nf = 5)
-##' acm$supv <- supcol(acm, dudi.acm(var_sup, scannf = FALSE, nf = 5)$tab)$cosup
+##' acm$supv <- supcol(acm, dudi.acm(var_sup, scannf = FALSE, nf = 5)$tab)
 ##' colw <- acm$cw*ncol(d)
 ##' X <- acm.disjonctif(ind_sup)
 ##' X <- data.frame(t(t(X)/colw) - 1)
-##' acm$supi <- suprow(acm, X)$lisup
+##' acm$supi <- suprow(acm, X)
 ##' explor(acm)
 ##' }
 
@@ -201,6 +201,8 @@ explor_multi_mca <- function(res, settings) {
                                 gettext("Labels size"),
                                 4, 20, 10),
                             explor_multi_auto_labels_input(res$vars, "var"),
+                            checkboxInput("var_prepend_varname", 
+                                gettext("Prepend variable name")),
                             sliderInput("var_point_size",
                                 gettext("Points size"),
                                 4, 128, 56),
@@ -214,6 +216,9 @@ explor_multi_mca <- function(res, settings) {
                                         "Supplementary variables"
                                     )),
                                     value = TRUE),
+                            conditionalPanel("input.var_sup",
+                                explor_multi_var_sup_choice_input(res$vars, settings)
+                            ),
                             explor_multi_sidebar_footer(type = "var")
                         )
                     ),
@@ -375,12 +380,15 @@ explor_multi_mca <- function(res, settings) {
                 size_range <- if (!is.null(input$var_size) && input$var_size != "None") c(30, 400) * input$var_point_size / 32 else c(10, 300)
                 var_lab_min_contrib <- if (settings$has_contrib) input$var_lab_min_contrib else 0
                 var_auto_labels <- if (!is.null(input$var_auto_labels) && input$var_auto_labels) "\"auto\"" else "NULL"
+                var_sup <- settings$has_sup_vars && input$var_sup
+                var_sup_choice <- if(var_sup) paste0(utils::capture.output(dput(input$var_sup_choice)), collapse="") else NULL
                 
                 paste0(
                     "explor::MCA_var_plot(res",
                     ", xax = ", input$var_x,
                     ", yax = ", input$var_y,
-                    ", var_sup = ", settings$has_sup_vars && input$var_sup,
+                    ", var_sup = ", var_sup,
+                    ", var_sup_choice = ", var_sup_choice,
                     ", var_lab_min_contrib = ", var_lab_min_contrib,
                     ", col_var = ", deparse(substitute(col_var)),
                     ", symbol_var = ", deparse(substitute(symbol_var)),
@@ -389,7 +397,8 @@ explor_multi_mca <- function(res, settings) {
                     ", labels_size = ", input$var_lab_size,
                     ", point_size = ", input$var_point_size,
                     ", transitions = ", input$var_transitions,
-                    ", labels_positions = ", var_auto_labels
+                    ", labels_positions = ", var_auto_labels,
+                    ", labels_prepend_var = ", input$var_prepend_varname
                 )
             })